backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.2
pydantic==2.11.1
pydantic-settings==2.8.1
python-multipart==0.0.20
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
argon2-cffi==23.1.0
